Instantiate project structure!  v0.0.1
Very bare project structure structure; provisional to change but set
out so that it is easy to expand on.

Within either cmd/ or tools/ I will be using different libraries, such
as the osm or dbpedia sub-modules.

I have presently made the decision to keep remote-specific submodules
in the internal directory, as I don't necessarily want them to be used
as libraries outside of this project.