Remove tools subdirectory!  v0.0.2
The tools subdirectory, as used in github.com/jakewilliami/tldinfo,
should be used for auxiliary programs/tools that aid the development
process.  At this stage, there is nothing that I need such a directory
for, as the primary function of the programme will be contained in the
cmd/ directory.  The tools directory's continued presence may prove
confusing in future; best to remove it now and add it back later if
needed.