Add basic provisional pull from remote!  v0.0.3
Since the 23rd of April, 2025, I have been playing around with the
various APIs I am to be interacting with (in particular, the Open
Street Maps (OSM) API).  This commit splits the work I have been doing
on probing these APIs and data into the submodules I have previously
defined, thereby instantiating very basic functionality to pull data
from remote sources and store it somewhere.

This is nowhere near complete, anvd very very messy, but as this is
work I have done in the previous probing project, this is a milestone,
and thus worthy of a provisional, pre-release.