Port current package functionality to Python!  v0.0.5
Port the basic concept/structure/probing of the library to Python, in
the interest of maintainability and access to Python's vast and mature
library ecosystem.

Porting the limited functionality I had for this package from Go to
Python was completed in early July, 2025, but then I took a hiatus
from this project while focussing on other work.  Coming back to it, I
wanted to finalise work on utility functions, with comprehensive
tests, which is why this was not completed till early August, 2025.